# packages/backend/src/modules/expenses/dto/create-expense.dto.ts
import { OmitType } from '@silte/nestjs-swagger';

import { ExpenseDto } from './expense.dto';

export class CreateExpenseDto extends OmitType(ExpenseDto, [
  'id',
  'userId',
  'v',
  'createdAt',
  'updatedAt',
] as const) {}