# packages/frontend/src/redux/schema/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts": {
      "post": {
        "operationId": "Accounts_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAccountDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "properties": { "payload": { "type": "string" } } }
              }
            }
          }
        },
        "tags": ["Accounts"]
      },
      "get": {
        "operationId": "Accounts_findAllByUser",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginationDto" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/AccountDto" }
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Accounts"]
      }
    },
    "/api/accounts/{id}": {
      "get": {
        "operationId": "Accounts_findOneById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return account by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountDto" }
              }
            }
          }
        },
        "tags": ["Accounts"]
      },
      "patch": {
        "operationId": "Accounts_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAccountDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AccountDto" }
              }
            }
          }
        },
        "tags": ["Accounts"]
      },
      "delete": {
        "operationId": "Accounts_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Accounts"]
      }
    },
    "/api/accounts/{id}/balance-history": {
      "get": {
        "operationId": "Accounts_getAccountBalanceHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return account balance history by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountBalanceHistoryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Accounts"]
      }
    },
...
    "/api/transactions/account/{id}": {
      "get": {
        "operationId": "Transactions_findAllByAccount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginationDto" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TransactionDto"
                          }
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Transactions"]
      }
    },
...
    "/api/expenses": {
      "get": {
        "operationId": "Expenses_findAllByUser",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/PaginationDto" },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/ExpenseDto" }
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Expenses"]
      },
      "post": {
        "operationId": "Expenses_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateExpenseDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseDto" }
              }
            }
          }
        },
        "tags": ["Expenses"]
      }
    },
    "/api/expenses/monthly-summaries": {
      "get": {
        "operationId": "Expenses_findMonthlySummariesByuser",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "accountTypes",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "transactionCategories",
            "required": false,
            "in": "query",
            "schema": { "type": "array", "items": { "type": "string" } }
          },
          {
            "name": "parentTransactionCategory",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionMonthSummaryDto"
                  }
                }
              }
            }
          }
        },
        "tags": ["Expenses"]
      }
    },
    "/api/expenses/{id}": {
      "get": {
        "operationId": "Expenses_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Return transaction by id",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseDto" }
              }
            }
          }
        },
        "tags": ["Expenses"]
      },
      "patch": {
        "operationId": "Expenses_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateExpenseDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ExpenseDto" }
              }
            }
          }
        },
        "tags": ["Expenses"]
      },
      "delete": {
        "operationId": "Expenses_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Expenses"]
      }
    },
...
  },
  "info": {
    "title": "Financer",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateTransactionCategoryMappingWithoutTransactionDto": {
        "type": "object",
        "properties": {
          "description": { "type": "string" },
          "categoryId": { "type": "string" },
          "amount": { "type": "number" }
        },
        "required": ["categoryId", "amount"]
      },
      "TransactionDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "fromAccount": { "type": "string" },
          "toAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "fromAccount",
          "toAccount",
          "categories"
        ]
      },
      "AccountType": {
        "type": "string",
        "enum": [
          "CASH",
          "SAVINGS",
          "INVESTMENT",
          "CREDIT",
          "LOAN",
          "LONG_TERM_SAVINGS",
          "PRE_ASSIGNED_CASH"
        ]
      },
      "AccountDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "name": { "type": "string" },
          "type": { "$ref": "#/components/schemas/AccountType" },
          "balance": { "type": "number" },
          "userId": { "type": "string" },
          "isDeleted": { "type": "boolean" }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "name",
          "type",
          "balance",
          "userId",
          "isDeleted"
        ]
      },
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "$ref": "#/components/schemas/AccountType" },
          "balance": { "type": "number" }
        },
        "required": ["name", "type", "balance"]
      },
      "PaginationDto": {
        "type": "object",
        "properties": {
          "currentPage": { "type": "number" },
          "totalPageCount": { "type": "number" },
          "totalRowCount": { "type": "number" },
          "limit": { "type": "number" },
          "hasPreviousPage": { "type": "boolean" },
          "hasNextPage": { "type": "boolean" }
        },
        "required": [
          "currentPage",
          "totalPageCount",
          "totalRowCount",
          "limit",
          "hasPreviousPage",
          "hasNextPage"
        ]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "type": { "$ref": "#/components/schemas/AccountType" },
          "balance": { "type": "number" }
        }
      },
      "AccountBalanceHistoryDto": {
        "type": "object",
        "properties": {
          "date": { "format": "date-time", "type": "string" },
          "amount": { "type": "number" },
          "balance": { "type": "number" }
        },
        "required": ["date", "amount", "balance"]
      },
      "TransactionMonthSummaryIdDto": {
        "type": "object",
        "properties": {
          "year": { "type": "number" },
          "month": { "type": "number" }
        },
        "required": ["year", "month"]
      },
      "TransactionMonthSummaryDto": {
        "type": "object",
        "properties": {
          "id": { "$ref": "#/components/schemas/TransactionMonthSummaryIdDto" },
          "count": { "type": "number" },
          "totalCount": { "type": "number" },
          "incomesCount": { "type": "number" },
          "expensesCount": { "type": "number" },
          "transferCount": { "type": "number" },
          "totalAmount": { "type": "number" },
          "incomeAmount": { "type": "number" },
          "expenseAmount": { "type": "number" },
          "transferAmount": { "type": "number" }
        },
        "required": [
          "id",
          "count",
          "totalCount",
          "incomesCount",
          "expensesCount",
          "transferCount",
          "totalAmount",
          "incomeAmount",
          "expenseAmount",
          "transferAmount"
        ]
      },
...
      "ExpenseDto": {
        "type": "object",
        "properties": {
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "id": { "type": "string" },
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "userId": { "type": "string" },
          "fromAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": [
          "createdAt",
          "updatedAt",
          "id",
          "amount",
          "description",
          "date",
          "userId",
          "fromAccount",
          "categories"
        ]
      },
      "CreateExpenseDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "fromAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        },
        "required": [
          "amount",
          "description",
          "date",
          "fromAccount",
          "categories"
        ]
      },
      "UpdateExpenseDto": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "description": { "type": "string" },
          "date": { "format": "date-time", "type": "string" },
          "fromAccount": { "type": "string" },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateTransactionCategoryMappingWithoutTransactionDto"
            }
          }
        }
      },
...
    }
  }
}